Create component for forgot password form
